requirements.txt
Django>=4.2
psycopg2-binary
asgiref
djangorestframework~=3.15.2
drf-yasg~=1.21.10
pandas~=2.2.3
django-phonenumber-field~=8.0.0
django-cors-headers
django-extensions
djangorestframework-simplejwt
django-redis
drf-yasg
phonenumbers
pillow
numpy
python-decouple~=3.8
dynaconf~=3.2.10
psycopg2-binary~=2.9
sqlparse
inflection
packaging>=21.0
pytz>=2021.1
uritemplate>=3.0.0
pyyaml
django-debug-toolbar